+++ pingora-core/src/protocols/tls/mod.rs
@@ -31,6 +31,7 @@ pub use rustls::*;
 
 #[cfg(not(feature = "any_tls"))]
 pub mod noop_tls;
+pub(crate) mod quic;
 
 #[cfg(not(feature = "any_tls"))]
 pub use noop_tls::*;
